feat: refactoring to use sqs queue
