+++ invalidate.env
@@ -1,1 +1,2 @@
 INVALIDATION_ENDPOINT=https://api.experiencecloud.live/publish-event/franklin-content-bus-headless
+SQS_QUEUE_URL=https://sqs.us-east-1.amazonaws.com/282898975672/franklin-json-ingestion.fifo
