+++ src/sqs-util.js
@@ -10,6 +10,12 @@
  * governing permissions and limitations under the License.
  */
 import { SQS } from '@aws-sdk/client-sqs';
+import processQueue from '@adobe/helix-shared-process-queue';
+import Storage from './storage.js';
+import { cloneObject, extractS3ObjectPath, sendSlackMessage } from './utils.js';
+import PullingClient from './pulling-client.js';
+import InvalidateClient from './invalidate-client.js';
+import VariationsUtil from './variations-util.js';
 
 export async function sendMessage(context, message) {
   const sqs = new SQS({
@@ -22,4 +28,62 @@ export async function sendMessage(context, message) {
     MessageGroupId: 'franklin-ingestor',
   };
   await sqs.sendMessage(params);
+}
+
+export async function processMessage(context, message) {
+  const {
+    action, mode, tenant, relPath, variation,
+  } = message;
+
+  const storage = new Storage(context);
+  const suffix = variation ? `/variations/${variation}` : '';
+  const s3ObjectPath = extractS3ObjectPath(message);
+
+  if (action === 'store') {
+    // init globalContext for given tenant
+    if (!context.globalContent[tenant]) {
+      try {
+        context.globalContent[tenant] = await storage.getKey(`${tenant}/settings.json`);
+      } catch (e) {
+        context.log.error(`Error while fetching settings for tenant ${tenant} due to ${e.message}`);
+      }
+    }
+    const settings = context.globalContent[tenant];
+    context.log.info(`Pulling content for ${relPath} in ${mode} mode`);
+    const data = await new PullingClient(
+      context,
+      settings[mode].baseURL,
+      settings[mode].authorization,
+    ).pullContent(relPath, variation);
+    if (data) {
+      context.log.info('Content pulled successfully storing in S3');
+      const storedKey = `${s3ObjectPath}.cfm.gql.json${suffix}`;
+      await storage.putKey(
+        storedKey,
+        data,
+        variation,
+      );
+      context.log.info(`putKey ${storedKey} success`);
+      await new InvalidateClient(context).invalidate(`${s3ObjectPath}.cfm.gql.json`, variation);
+      if (!variation) {
+        const variationMessages = await new VariationsUtil(
+          context,
+          message,
+        ).process(data, message);
+        processQueue(cloneObject(variationMessages), async (variationMessage) => {
+          await processMessage(context, variationMessage);
+        });
+      }
+      const varSelector = variation ? `${variation}.` : '';
+      const varMessage = variation ? ` for variation ${variation}` : '';
+      const url = `${settings[mode].external}/content/dam/${relPath}.cfm.gql.${varSelector}json`;
+      await sendSlackMessage(context.globalContent[tenant], `Fully hydrated json <${url}|${relPath}> is ready to view in ${mode} mode${varMessage}`);
+    }
+  } else {
+    const evictedKeys = [];
+    evictedKeys.push(...await storage.evictKeys(s3ObjectPath));
+    await new InvalidateClient(context).invalidateAll(
+      evictedKeys,
+    );
+  }
 }
